+++ apps/codeimage/src/components/Presets/PresetSwitcher/PresetSwitcher.tsx
@@ -5,7 +5,7 @@ import {getTerminalState} from '@codeimage/store/editor/terminal';
 import {getPresetsStore} from '@codeimage/store/presets/presets';
 import {PresetData} from '@codeimage/store/presets/types';
 import {getUiStore} from '@codeimage/store/ui';
-import {Box, HStack, Text} from '@codeimage/ui';
+import {Box, HStack, Text, toast} from '@codeimage/ui';
 import {
   Button,
   DropdownMenu,
@@ -15,7 +15,6 @@ import {
   DropdownMenuTrigger,
   IconButton,
   Tooltip,
-  As,
 } from '@codeui/kit';
 import {getUmami} from '@core/constants/umami';
 import {formatDistanceToNow} from '@core/helpers/date';
@@ -176,15 +175,13 @@ export const PresetSwitcher: ParentComponent<
                           }}
                         >
                           <DropdownMenu>
-                            <DropdownMenuTrigger asChild>
-                              <As
-                                component={IconButton}
-                                aria-label={'Menu'}
-                                theme={'secondary'}
-                                size={'xs'}
-                              >
-                                <DotHorizontalIcon size={'md'} />
-                              </As>
+                            <DropdownMenuTrigger
+                              as={IconButton}
+                              aria-label={'Menu'}
+                              theme={'secondary'}
+                              size={'xs'}
+                            >
+                              <DotHorizontalIcon size={'md'} />
                             </DropdownMenuTrigger>
                             <DropdownMenuPortal>
                               <DropdownMenuContent>
@@ -213,6 +210,16 @@ export const PresetSwitcher: ParentComponent<
                                   }}
                                 >
                                   {t('presets.renamePreset.label')}
+                                </DropdownMenuItem>
+                                <DropdownMenuItem
+                                  onClick={() => {
+                                    presetsStore.actions.copyLink(theme);
+                                    toast.success(t('presets.share.confirm'), {
+                                      position: 'bottom-center',
+                                    });
+                                  }}
+                                >
+                                  {t('presets.share.label')}
                                 </DropdownMenuItem>
                                 <DropdownMenuItem
                                   onClick={() => {
@@ -245,7 +252,7 @@ export const PresetSwitcher: ParentComponent<
                           justifyContent={'flexEnd'}
                           paddingTop={3}
                         >
-                          <Button
+                          <Button<'button'>
                             theme={'secondary'}
                             block
                             size={'sm'}