+++ apps/codeimage/src/pages/NotFound/NotFoundPage.tsx
@@ -1,6 +1,6 @@
 import {useI18n} from '@codeimage/locale';
 import {Box, Text} from '@codeimage/ui';
-import {As, Button} from '@codeui/kit';
+import {Button} from '@codeui/kit';
 import {Link} from '@solidjs/router';
 import {AppLocaleEntries} from '../../i18n';
 import * as styles from './NotFoundPage.css';
@@ -13,11 +13,15 @@ export default function NotFoundPage() {
       <Text class={styles.notFoundTitle}>{t('notFound.title')}</Text>
       <Text class={styles.descriptionTitle}>{t('notFound.description')}</Text>
       <Box marginTop={12}>
-        <Button asChild size={'md'} theme={'primary'}>
+        <Button
+          size={'md'}
+          theme={'primary'}
+          as={Link}
+          href={'/'}
+          style={{'text-decoration': 'unset'}}
+        >
           {/*// TODO: remove text decoration add link*/}
-          <As component={Link} href={'/'} style={{'text-decoration': 'unset'}}>
-            {t('notFound.goToHome')}
-          </As>
+          {t('notFound.goToHome')}
         </Button>
       </Box>
     </div>