+++ apps/codeimage/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @codeimage/app
 
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies [[`9bc923d7`](https://github.com/riccardoperra/codeimage/commit/9bc923d7bbfc6e41ae3aa5a3b034e731d1107ea4)]:
+  - @codeimage/config@9.0.2
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies [[`e5a3ec31`](https://github.com/riccardoperra/codeimage/commit/e5a3ec31547650031536980545475b4d3e27906b)]:
+  - @codeimage/config@9.0.1
+
+## 1.8.0
+
+### Minor Changes
+
+- [#647](https://github.com/riccardoperra/codeimage/pull/647) [`db76eb05`](https://github.com/riccardoperra/codeimage/commit/db76eb0564534953ed3c05d1daadfc680d8277ae) Thanks [@riccardoperra](https://github.com/riccardoperra)! - Update @codeui dependency, update vite/rollup and infra deps
+
+### Patch Changes
+
+- Updated dependencies [[`db76eb05`](https://github.com/riccardoperra/codeimage/commit/db76eb0564534953ed3c05d1daadfc680d8277ae)]:
+  - @codeimage/highlight@0.11.0
+  - @codeimage/locale@0.7.0
+  - @codeimage/api@0.5.0
+  - @codeimage/config@9.0.0
+  - @codeimage/ui@0.10.5
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies [[`581ee890`](https://github.com/riccardoperra/codeimage/commit/581ee8905775d0a0d37cf1077a89b925afa29f16)]:
+  - @codeimage/config@8.2.8
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies [[`5bb5f2a8`](https://github.com/riccardoperra/codeimage/commit/5bb5f2a824e69d7cd3d9c2616eaed2409dce5626)]:
+  - @codeimage/highlight@0.10.10
+  - @codeimage/config@8.2.7
+
+## 1.7.0
+
+### Minor Changes
+
+- [#637](https://github.com/riccardoperra/codeimage/pull/637) [`06f6e624`](https://github.com/riccardoperra/codeimage/commit/06f6e624e2510e585ca5ce0cce79b794eb0f81dd) Thanks [@riccardoperra](https://github.com/riccardoperra)! - Add shareable presets
+
 ## 1.6.2
 
 ### Patch Changes