+++ packages/highlight/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @codeimage/highlight
 
+## 0.11.0
+
+### Minor Changes
+
+- [#647](https://github.com/riccardoperra/codeimage/pull/647) [`db76eb05`](https://github.com/riccardoperra/codeimage/commit/db76eb0564534953ed3c05d1daadfc680d8277ae) Thanks [@riccardoperra](https://github.com/riccardoperra)! - Update @codeui dependency, update vite/rollup and infra deps
+
+## 0.10.10
+
+### Patch Changes
+
+- [#642](https://github.com/riccardoperra/codeimage/pull/642) [`5bb5f2a8`](https://github.com/riccardoperra/codeimage/commit/5bb5f2a824e69d7cd3d9c2616eaed2409dce5626) Thanks [@riccardoperra](https://github.com/riccardoperra)! - fix: update fleet theme
+
 ## 0.10.9
 
 ### Patch Changes